+++ react/index.tsx
@@ -1,6 +1,6 @@
 import React, { Fragment, FC, useRef, useEffect, useCallback } from 'react'
 import { generateBlockClass, BlockClass } from '@vtex/css-handles'
-import { path } from 'ramda'
+import { path, head } from 'ramda'
 
 import styles from './styles.css'
 
@@ -14,6 +14,15 @@ interface Props {
 
 interface StorefrontComponent extends FC<Props & BlockClass> {
   schema?: any
+}
+
+const transformPxToInt = (value?: string) => {
+  if (!value) {
+    return 0
+  }
+  const isNegative = head(value) === '-'
+  const number = parseInt(value.replace(/\D/g, ''), 10)
+  return isNegative ? -1 * number : number
 }
 
 const StickyLayoutComponent: StorefrontComponent = ({
@@ -30,16 +39,24 @@ const StickyLayoutComponent: StorefrontComponent = ({
     return <Fragment>{children}</Fragment>
   }
 
-  const handleScroll = useCallback(() => {
+  const handleScroll = useCallback((e?: Event) => {
+    const target = e && (e.target as HTMLDivElement | null)
     if (!container.current) {
       return
     }
     if (startTop.current == null) {
       startTop.current = container.current.offsetTop
     }
+
+    // We have to get this baseTop variable to handle cases when a modal changes the top style of the base element, resetting the window.pageYOfsset
+    const baseTop = transformPxToInt(
+      path(['scrollingElement', 'style', 'top'], target)
+    )
+    const componentTop = startTop.current + baseTop
+
     const newTop =
       window.innerHeight -
-      startTop.current -
+      componentTop -
       container.current.clientHeight +
       window.pageYOffset